pyproject.toml
[project]
name = "langgraph-agents-toys"
version = "0.1.0"
description = ""
authors = [
    {name = "Saite Guo",email = "131952679+Web-commited@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "streamlit (>=1.42.0,<2.0.0)",
    "langgraph (>=0.2.70,<0.3.0)",
    "boto3 (>=1.36.15,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.17,<0.4.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "pygraphviz (>=1.14,<2.0)",
    "langchain-aws (>=0.2.12,<0.3.0)",
    "ipython (>=8.32.0,<9.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "duckduckgo-search (>=7.3.1,<8.0.0)",
    "tavily-python (>=0.5.1,<0.6.0)",
    "defusedxml (>=0.7.1,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
